refactor: restructure tags model and schema for improved organization
- Removed the old tags.model.ts file and integrated its content into tags.schema.ts for better cohesion.
- Updated the schema definitions for tags and article_tags, ensuring proper relationships and references.
- Enhanced the organization of import statements to maintain consistency across the tags module.
