+++ src/tags/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './tag-response.dto';
+export * from './article-tag.dto';
