+++ src/tags/interfaces/tag-row.interface.ts
@@ -1,0 +1,4 @@
+import type { InferSelectModel } from 'drizzle-orm';
+import type { tags } from '../tags.schema';
+
+export type TagRow = InferSelectModel<typeof tags>;
