+++ src/tags/interfaces/tag.interface.ts
@@ -1,0 +1,5 @@
+export type ITag = {
+  name: string;
+  createdAt: Date;
+  updatedAt: Date;
+};
